+++ changelog.md
@@ -1,3 +1,64 @@
+# 0.4.4
+
+### New Features
+* LSL streaming example (thanks @gabrielibagon)
+
+### Bug Fixes
+* Fix #27 by adding delay
+
+# 0.4.3
+
+### Bug Fixes
+* Fix #23 by hard setting noble to 0.1.7, 0.1.8 breaks
+
+# 0.4.2
+
+### New Features
+* Add callback function to constructor to catch noble errors.
+
+### Bug Fixes
+* Fix #19
+
+# 0.4.1
+
+### New Features
+* Dropped connection will fire disconnect
+
+### Bug Fixes
+* Improvements to disconnect
+
+# 0.4.0
+
+### New Features
+* If dropped connection, module will now disconnect and clean up.
+
+### Breaking Changes
+* Dropped connection will fire disconnect
+* Auto reconnect functionality temporarily removed.
+
+# 0.3.8
+
+### New Features
+* Can now subscribe to `scanStart` and `scanStop` to better control scanning.
+
+# 0.3.7
+
+### Bug Fixes
+* Improvements to disconnect
+
+# 0.3.6
+
+### Bug Fixes
+* Fixed bug where not disconnecting.
+
+# 0.3.5 - 0.3.4 - 0.3.3
+
+### New Features
+* Get state of noble with `.isNobleReady()`
+
+### Bug Fixes
+* Fix bug where scan stop would not stop scan which led to many problems.
+
 # 0.3.2
 
 ### Bug Fixes